requirements_base.txt
Flask
Flask-SQLAlchemy
flask-marshmallow
marshmallow-sqlalchemy
Flask-JWT-Extended
Flask-Mail
Flask-WTF
python-dotenv
python-slugify
requests
